chore: rename handlers into publicHandlers
